refactor(daily-goal): remove unused API functions and simplify state management
This commit deletes the `addDailyGoal`, `fetchDailyGoals`, `updDailyGoal`, and related exports, which were no longer in use. It also consolidates the logic for updating and fetching daily goals into the `saveDailyGoal` and `fetchAndSetDailyGoal` functions in the utils file. This change simplifies the codebase and improves maintainability by reducing redundancy and focusing on the core functionalities required for handling daily goals.
