--- src/features/daily-goal/api/index.ts
@@ -1,3 +1,0 @@
-export * from './addDailyGoal'
-export * from './fetchDailyGoals'
-export * from './updDailyGoal'
