+++ src/features/daily-goal/utils.ts
@@ -1,6 +1,16 @@
-import { DailyGoal } from '../../API'
+import {
+  CreateDailyGoalInput,
+  CreateDailyGoalMutationVariables,
+  DailyGoal,
+  UpdateDailyGoalInput,
+  UpdateDailyGoalMutationVariables,
+} from '../../API'
+import {
+  addDailyGoal,
+  fetchDailyGoal,
+  updDailyGoal,
+} from '../../api/daily-goal'
 import { DailyGoalState } from '../../stores'
-import { addDailyGoal, updDailyGoal } from './api'
 
 /**
  * Create initial values for a daily goal.
@@ -25,9 +35,10 @@ export const createDailyGoalInitialValues = (): DailyGoal => {
 }
 
 /**
- * Saves the daily goal by either updating an existing goal or adding a new one.
+ * Saves the daily goal by either updating an existing goal or creating a new one.
  *
- * @param dailyGoal - The daily goal object to be saved. It contains the details of the daily goal.
+ * @param dailyGoal - The current state of the daily goal.
+ * @param setDailyGoal - A function to update the state of the daily goal.
  * @returns A promise that resolves when the daily goal has been saved.
  */
 export const saveDailyGoal = async (
@@ -37,10 +48,49 @@ export const saveDailyGoal = async (
   const dailyGoalId = dailyGoal.id
 
   if (dailyGoalId) {
-    await updDailyGoal(dailyGoal, setDailyGoal)
+    const updateDailyGoalInput: UpdateDailyGoalInput = {
+      id: dailyGoal.id,
+      calories: dailyGoal.calories,
+      protein: dailyGoal.protein,
+      carbohydrates: dailyGoal.carbohydrates,
+      fat: dailyGoal.fat,
+      _version: dailyGoal._version,
+    }
+
+    const variables: UpdateDailyGoalMutationVariables = {
+      input: updateDailyGoalInput,
+    }
+
+    const updatedDailyGoal = await updDailyGoal(variables)
+    setDailyGoal(updatedDailyGoal as DailyGoal)
   }
 
   if (!dailyGoalId) {
-    await addDailyGoal(dailyGoal, setDailyGoal)
+    const createDailyGoalInput: CreateDailyGoalInput = {
+      calories: dailyGoal.calories,
+      protein: dailyGoal.protein,
+      carbohydrates: dailyGoal.carbohydrates,
+      fat: dailyGoal.fat,
+    }
+
+    const variables: CreateDailyGoalMutationVariables = {
+      input: createDailyGoalInput,
+    }
+
+    const newDailyGoal = await addDailyGoal(variables)
+    setDailyGoal(newDailyGoal as DailyGoal)
   }
+}
+
+/**
+ * Fetches the daily goal from the server and sets it in the state.
+ *
+ * @param setDailyGoal - A function to update the daily goal state.
+ * @returns A promise that resolves when the daily goal has been fetched and set.
+ */
+export const fetchAndSetDailyGoal = async (
+  setDailyGoal: DailyGoalState['setDailyGoal']
+) => {
+  const dailyGoal = await fetchDailyGoal()
+  setDailyGoal(dailyGoal as DailyGoal)
 }
